--- app/Http/Controllers/CarritoController.php
@@ -1,66 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Http\Requests\StoreCarritoRequest;
-use App\Http\Requests\UpdateCarritoRequest;
-use App\Models\Carrito;
-
-class CarritoController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(StoreCarritoRequest $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Carrito $carrito)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Carrito $carrito)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateCarritoRequest $request, Carrito $carrito)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Carrito $carrito)
-    {
-        //
-    }
-}
