--- app/Http/Controllers/CarritoDetalleController.php
@@ -1,66 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Http\Requests\StoreCarritoDetalleRequest;
-use App\Http\Requests\UpdateCarritoDetalleRequest;
-use App\Models\CarritoDetalle;
-
-class CarritoDetalleController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(StoreCarritoDetalleRequest $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(CarritoDetalle $carritoDetalle)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(CarritoDetalle $carritoDetalle)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateCarritoDetalleRequest $request, CarritoDetalle $carritoDetalle)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(CarritoDetalle $carritoDetalle)
-    {
-        //
-    }
-}
