--- app/Models/Carrito.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
-
-class Carrito extends Model
-{
-    /** @use HasFactory<\Database\Factories\CarritoFactory> */
-    use HasFactory, SoftDeletes;
-
-    protected $fillable = ['user_id'];
-
-    public function user()
-    {
-        return $this->belongsTo(User::class);
-    }
-
-    public function detalles()
-    {
-        return $this->hasMany(CarritoDetalle::class);
-    }
-}
