--- app/Policies/CarritoDetallePolicy.php
@@ -1,66 +1,0 @@
-<?php
-
-namespace App\Policies;
-
-use App\Models\CarritoDetalle;
-use App\Models\User;
-use Illuminate\Auth\Access\Response;
-
-class CarritoDetallePolicy
-{
-    /**
-     * Determine whether the user can view any models.
-     */
-    public function viewAny(User $user): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can view the model.
-     */
-    public function view(User $user, CarritoDetalle $carritoDetalle): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can create models.
-     */
-    public function create(User $user): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can update the model.
-     */
-    public function update(User $user, CarritoDetalle $carritoDetalle): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can delete the model.
-     */
-    public function delete(User $user, CarritoDetalle $carritoDetalle): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can restore the model.
-     */
-    public function restore(User $user, CarritoDetalle $carritoDetalle): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can permanently delete the model.
-     */
-    public function forceDelete(User $user, CarritoDetalle $carritoDetalle): bool
-    {
-        return false;
-    }
-}
